requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
asgiref
asttokens
attrs
authy
bcrypt
beautifulsoup4
better-profanity
bleach
boto3
botocore
certifi
cffi
charset-normalizer
colorama
comm
crispy-bootstrap4
cryptography
debugpy
decorator
Django
django-cache-machine
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-ratelimit
django-storages
djangorestframework
djangorestframework-simplejwt
djongo
dnspython
executing
frozenlist
google
google-cloud
gunicorn
idna
ipykernel
ipython
jedi
jmespath
jupyter_client
jupyter_core
matplotlib-inline
multidict
mysqlclient
ndg-httpsclient
nest-asyncio
notification
numpy
packaging
parso
pillow
platformdirs
prompt_toolkit
psutil
psycopg2-binary
pure_eval
pusher
pyasn1
pycparser
Pygments
PyJWT
PyMySQL
PyNaCl
pyOpenSSL
python-dateutil
python-decouple
python-dotenv
python-memcached
pytz
# pywin32  # Commented out or specify a valid version
pyzmq
requests
s3transfer
six
soupsieve
sqlparse
stack-data
threadpoolctl
tokenize-rt
tornado
traitlets
twilio
typing_extensions
tzdata
uritemplate
urllib3
wcwidth
webencodings
wheel
whitenoise
wrapt
yarl